Refatora Storage's
