+++ src/E5R.Architecture.Data/Abstractions/IBulkStorageWriter.cs
@@ -1,14 +1,6 @@
-﻿using System.Collections.Generic;
-
-namespace E5R.Architecture.Data.Abstractions
+﻿namespace E5R.Architecture.Data.Abstractions
 {
-    public interface IBulkStorageWriter<TModel, TIdentifier>
-        where TModel : DataModel<TIdentifier>
-        where TIdentifier : struct
+    public interface IBulkStorageWriter : IBulkStorageWriter<DataModel, VoidIdentifier>
     {
-        IEnumerable<TModel> BulkCreate(IEnumerable<TModel> data);
-        IEnumerable<TModel> BulkReplace(IEnumerable<TModel> models);
-        void BulkRemove(IEnumerable<TIdentifier> ids);
-        void BulkRemoveFromSearch(DataReducer<TModel, TIdentifier> reducer);
     }
 }
