+++ src/E5R.Architecture.Data/Abstractions/IBulkStorageWriter`2.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+
+namespace E5R.Architecture.Data.Abstractions
+{
+    public interface IBulkStorageWriter<TModel, TIdentifier>
+        where TModel : DataModel<TIdentifier>
+        where TIdentifier : struct
+    {
+        IEnumerable<TModel> BulkCreate(IEnumerable<TModel> data);
+        IEnumerable<TModel> BulkReplace(IEnumerable<TModel> models);
+        void BulkRemove(IEnumerable<TIdentifier> ids);
+        void BulkRemoveFromSearch(DataReducer<TModel, TIdentifier> reducer);
+    }
+}
