+++ src/E5R.Architecture.Data/Abstractions/IStorage.cs
@@ -1,11 +1,6 @@
 ﻿namespace E5R.Architecture.Data.Abstractions
 {
-    public interface IStorage<TModel, TIdentifier> :
-        IStorageReader<TModel, TIdentifier>,
-        IStorageWriter<TModel, TIdentifier>,
-        IBulkStorageWriter<TModel, TIdentifier>
-        where TModel : DataModel<TIdentifier>
-        where TIdentifier : struct
+    public interface IStorage : IStorageReader, IStorageWriter, IBulkStorageWriter
     {
     }
 }
