+++ src/E5R.Architecture.Data/Abstractions/IStorageWriter.cs
@@ -1,11 +1,6 @@
 ﻿namespace E5R.Architecture.Data.Abstractions
 {
-    public interface IStorageWriter<TModel, in TIdentifier>
-        where TModel : DataModel<TIdentifier>
-        where TIdentifier : struct
+    public interface IStorageWriter : IStorageWriter<DataModel, VoidIdentifier>
     {
-        TModel Create(TModel data);
-        TModel Replace(TModel data);
-        void Remove(TIdentifier id);
     }
 }
