+++ src/E5R.Architecture.Data/Abstractions/IStorageWriter`2.cs
@@ -1,0 +1,11 @@
+﻿namespace E5R.Architecture.Data.Abstractions
+{
+    public interface IStorageWriter<TModel, in TIdentifier>
+        where TModel : DataModel<TIdentifier>
+        where TIdentifier : struct
+    {
+        TModel Create(TModel data);
+        TModel Replace(TModel data);
+        void Remove(TIdentifier id);
+    }
+}
