+++ src/E5R.Architecture.Data/Abstractions/IStorage`2.cs
@@ -1,0 +1,11 @@
+﻿namespace E5R.Architecture.Data.Abstractions
+{
+    public interface IStorage<TModel, TIdentifier> :
+        IStorageReader<TModel, TIdentifier>,
+        IStorageWriter<TModel, TIdentifier>,
+        IBulkStorageWriter<TModel, TIdentifier>
+        where TModel : DataModel<TIdentifier>
+        where TIdentifier : struct
+    {
+    }
+}
