+++ src/E5R.Architecture.Data/E5R.Architecture.Data.csproj
@@ -1,12 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <Import Project="..\..\packages.props"/>
-
+  <Import Project="..\..\packages.props" />
   <PropertyGroup>
     <AssemblyName>E5R.Architecture.Data</AssemblyName>
     <PackageId>E5R.Architecture.Data</PackageId>
     <RootNamespace>$(RootPkgNamespace).Data</RootNamespace>
     <Description>Componente de dados, parte da arquitetura de software E5R Development Team</Description>
   </PropertyGroup>
-
 </Project>
